+++ include/crocoddyl/core/codegen/action.hxx
@@ -2,7 +2,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 // BSD 3-Clause License
 //
-// Copyright (C) 2019-2025, LAAS-CNRS, INRIA, University of Edinburgh,
+// Copyright (C) 2019-2026, LAAS-CNRS, INRIA, University of Edinburgh,
 //                          Heriot-Watt University
 // Copyright note valid unless otherwise stated in individual files.
 // All rights reserved.
@@ -365,11 +365,23 @@ void ActionModelCodeGenTpl<Scalar>::compileLib() {
                  << lib_fname_ + SystemInfo::DYNAMIC_LIB_EXTENSION
                  << " should not be compiled again");
   }
+  auto splitFlags = [](const std::string& flags) {
+    std::istringstream iss(flags);
+    std::vector<std::string> out;
+    std::string flag;
+    while (iss >> flag) {
+      out.push_back(flag);
+    }
+    return out;
+  };
   switch (compiler_type_) {
     case GCC: {
       CppAD::cg::GccCompiler<Scalar> compiler("/usr/bin/gcc");
       std::vector<std::string> compile_flags = compiler.getCompileFlags();
-      compile_flags[0] = compile_options_;
+      compile_flags.clear();
+      auto extra_flags = splitFlags(compile_options_);
+      compile_flags.insert(compile_flags.end(), extra_flags.begin(),
+                           extra_flags.end());
       compiler.setCompileFlags(compile_flags);
       dynLibManager_->createDynamicLibrary(compiler, false);
       break;
@@ -377,7 +389,10 @@ void ActionModelCodeGenTpl<Scalar>::compileLib() {
     case CLANG: {
       CppAD::cg::ClangCompiler<Scalar> compiler("/usr/bin/clang");
       std::vector<std::string> compile_flags = compiler.getCompileFlags();
-      compile_flags[0] = compile_options_;
+      compile_flags.clear();
+      auto extra_flags = splitFlags(compile_options_);
+      compile_flags.insert(compile_flags.end(), extra_flags.begin(),
+                           extra_flags.end());
       compiler.setCompileFlags(compile_flags);
       dynLibManager_->createDynamicLibrary(compiler, false);
       break;
@@ -651,7 +666,7 @@ ActionModelCodeGenTpl<Scalar>::ActionModelCodeGenTpl()
       np_(0),
       lib_fname_(""),
       compiler_type_(CLANG),
-      compile_options_("-Ofast -march=native"),
+      compile_options_("-O -ffast-math -march=native"),
       updateParams_(EmptyParamsEnv) {
   // Add initialization logic if necessary
 }